requirements.txt
python-barcode
pdfkit~=0.6.1